Register jobs module and add a navigation entry
Register the new module to make the jobs endpoint available in Alchemy. Add a new jobs controller and secure them with a custom ability which is only granted for admins. This way it is possible to disable the http authentication of MissionControl::Jobs and use the Jobs controller as base controller. This way the mission control backend is secured by Alchemy.