Adjust the background color of mission control iframe
Use light grey background color for mission control backend. This way the iframe does not look like an iframe anymore. It is still pretty odd, because the UI differs from Alchemy, but it is good enough for the first implementation. Also the HTML was a bit adjusted to reduce the padding.